+++ src/hardis-colors.ts
@@ -6,12 +6,14 @@ import {
   hasSfdxProjectJson,
   listLocalSfConfigFilePaths,
   listLocalSfConfigFiles,
-  loadFromLocalConfigFile,
-  readSfdxHardisConfig,
-  writeSfdxHardisConfig,
 } from "./utils";
 import { HardisStatusProvider } from "./hardis-status-provider";
 import { t } from "./i18n/i18n";
+import {
+  loadFromLocalConfigFile,
+  readSfdxHardisConfig,
+  writeSfdxHardisConfig,
+} from "./utils/sfdx-hardis-config-utils";
 
 const PRODUCTION_EDITIONS = [
   "Team Edition",